pyproject.toml
[project]
name = "ultools-comfyui"
description = "Nodes:SaveImgAdv, CLIPTextEncodeWithStats, OpenPoseEditorAdv, SolidMaskAdv. Collection of tools supporting txt2img generation in ComfyUI and other tasks."
version = "1.0.0"
license = "LICENSE"
dependencies = ["pandas", "numpy", "piexif", "pillow"]

[project.urls]
Repository = "https://github.com/jkrauss82/ultools-comfyui"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "uberlord"
DisplayName = "ultools-comfyui"
Icon = ""
